HSEARCH-3289 Remove any mention of "stream" workers/workplans
Note we remove some operations (flush, optimize), but they will be
re-introduced as part of different APIs.

See https://hibernate.atlassian.net/browse/HSEARCH-3305.
