+++ backend-lucene/src/main/java/org/hibernate/search/v6poc/backend/lucene/index/impl/IndexingBackendContext.java
@@ -9,7 +9,6 @@ package org.hibernate.search.v6poc.backend.lucene.index.impl;
 import java.io.IOException;
 
 import org.hibernate.search.v6poc.backend.index.spi.ChangesetIndexWorker;
-import org.hibernate.search.v6poc.backend.index.spi.StreamIndexWorker;
 import org.hibernate.search.v6poc.backend.lucene.document.impl.LuceneRootDocumentBuilder;
 import org.hibernate.search.v6poc.backend.lucene.multitenancy.impl.MultiTenancyStrategy;
 import org.hibernate.search.v6poc.backend.lucene.orchestration.impl.LuceneIndexWorkOrchestrator;
@@ -55,15 +54,6 @@ public class IndexingBackendContext {
 		multiTenancyStrategy.checkTenantId( sessionContext.getTenantIdentifier(), eventContext );
 
 		return new LuceneChangesetIndexWorker( workFactory, multiTenancyStrategy, orchestrator,
-				indexName, sessionContext );
-	}
-
-	StreamIndexWorker<LuceneRootDocumentBuilder> createStreamIndexWorker(
-			LuceneIndexWorkOrchestrator orchestrator,
-			String indexName, SessionContext sessionContext) {
-		multiTenancyStrategy.checkTenantId( sessionContext.getTenantIdentifier(), eventContext );
-
-		return new LuceneStreamIndexWorker( workFactory, multiTenancyStrategy, orchestrator,
 				indexName, sessionContext );
 	}
 }
