+++ documentation/src/main/asciidoc/internals.asciidoc
@@ -147,20 +147,16 @@ The accessor is responsible for exactly this: given a document,
 it sets the value of the field it represents to some mapper-provided value.
 
 The other part of indexing (or altering the index in any way) is to give an order to the index manager:
-"add this document", "delete this document", "flush the index", ...
-This is done through the index "workers": `ChangesetIndexWorker` and `StreamIndexWorker`.
+"add this document", "delete this document", ...
+This is done through the `ChangesetIndexWorker` class.
 The mapper should create a worker whenever it needs to execute a series of works.
 
-Which type of worker to create depends on the use case:
-
-* When executing a relatively short, complex sequence of potentially related works,
-`ChangesetIndexWorker` should be used. Implementations may not be thread safe.
-* When executing lots of unrelated, simple works (for instance when reindexing a large number of objects),
-`StreamIndexWorker` should be used. Implementations are thread safe.
-
-Workers carry *some* context usually associated to a "session" in the JPA world,
+`ChangesetIndexWorker` carries *some* context usually associated to a "session" in the JPA world,
 including the tenant identifier when using multi-tenancy, in particular.
 Thus the mapper should instantiate a new worker whenever this context changes.
+
+NOTE: For now index-scoped operations such as flush, optimize, etc. are unavailable from workers.
+HSEARCH-3305 will introduce APIs and SPIs for these.
 
 ==== Searching
