+++ mapper-pojo/src/main/java/org/hibernate/search/v6poc/entity/pojo/mapping/ChangesetPojoWorker.java
@@ -9,7 +9,9 @@ package org.hibernate.search.v6poc.entity.pojo.mapping;
 import java.util.concurrent.CompletableFuture;
 
 /**
- * A worker that accumulates works in a list (called a changeset),
+ * An entry point to execute works on POJO-mapped indexes.
+ * <p>
+ * Accumulates works in a list (called a changeset),
  * and executes them only when {@link #execute()} is called.
  * <p>
  * Relative ordering of works within a changeset will be preserved.
@@ -18,7 +20,47 @@ import java.util.concurrent.CompletableFuture;
  *
  * @author Yoann Rodiere
  */
-public interface ChangesetPojoWorker extends PojoWorker {
+public interface ChangesetPojoWorker {
+
+	/**
+	 * Add an entity to the index, assuming that the entity is absent from the index.
+	 * <p>
+	 * Shorthand for {@code add(null, entity)}; see {@link #add(Object, Object)}.
+	 *
+	 * @param entity The entity to add to the index.
+	 */
+	void add(Object entity);
+
+	/**
+	 * Add an entity to the index, assuming that the entity is absent from the index.
+	 * <p>
+	 * <strong>Note:</strong> depending on the backend, this may lead to errors or duplicate entries in the index
+	 * if the entity was actually already present in the index before this call.
+	 * When in doubt, you should rather use {@link #update(Object, Object)} or {@link #update(Object)}.
+	 *
+	 * @param id The provided ID for the entity.
+	 * If {@code null}, Hibernate Search will attempt to extract the ID from the entity.
+	 * @param entity The entity to add to the index.
+	 */
+	void add(Object id, Object entity);
+
+	/**
+	 * Update an entity in the index, or add it if it's absent from the index.
+	 * <p>
+	 * Shorthand for {@code update(null, entity)}; see {@link #update(Object, Object)}.
+	 *
+	 * @param entity The entity to update in the index.
+	 */
+	void update(Object entity);
+
+	/**
+	 * Update an entity in the index, or add it if it's absent from the index.
+	 *
+	 * @param id The provided ID for the entity.
+	 * If {@code null}, Hibernate Search will attempt to extract the ID from the entity.
+	 * @param entity The entity to update in the index.
+	 */
+	void update(Object id, Object entity);
 
 	/**
 	 * Update an entity in the index, or add it if it's absent from the index,
@@ -47,6 +89,26 @@ public interface ChangesetPojoWorker extends PojoWorker {
 	 * ("directEntityProperty.nestedPropery").
 	 */
 	void update(Object id, Object entity, String... dirtyPaths);
+
+	/**
+	 * Delete an entity from the index.
+	 * <p>
+	 * Shorthand for {@code delete(null, entity)}; see {@link #delete(Object, Object)}.
+	 *
+	 * @param entity The entity to delete from the index.
+	 */
+	void delete(Object entity);
+
+	/**
+	 * Delete an entity from the index.
+	 * <p>
+	 * No effect on the index if the entity is not in the index.
+	 *
+	 * @param id The provided ID for the entity.
+	 * If {@code null}, Hibernate Search will attempt to extract the ID from the entity.
+	 * @param entity The entity to delete from the index.
+	 */
+	void delete(Object id, Object entity);
 
 	/**
 	 * Prepare the changeset execution, i.e. execute as much as possible without writing to the index.
