+++ mapper-pojo/src/main/java/org/hibernate/search/v6poc/entity/pojo/mapping/impl/PojoMappingDelegateImpl.java
@@ -14,7 +14,6 @@ import java.util.Set;
 import org.hibernate.search.v6poc.engine.spi.SessionContext;
 import org.hibernate.search.v6poc.entity.pojo.logging.impl.Log;
 import org.hibernate.search.v6poc.entity.pojo.mapping.ChangesetPojoWorker;
-import org.hibernate.search.v6poc.entity.pojo.mapping.StreamPojoWorker;
 import org.hibernate.search.v6poc.entity.pojo.mapping.spi.PojoMappingDelegate;
 import org.hibernate.search.v6poc.entity.pojo.mapping.spi.PojoSearchTargetDelegate;
 import org.hibernate.search.v6poc.entity.pojo.mapping.spi.PojoSessionContext;
@@ -50,11 +49,6 @@ public class PojoMappingDelegateImpl implements PojoMappingDelegate {
 	@Override
 	public ChangesetPojoWorker createWorker(PojoSessionContext sessionContext) {
 		return new ChangesetPojoWorkerImpl( indexedTypeManagers, containedTypeManagers, sessionContext );
-	}
-
-	@Override
-	public StreamPojoWorker createStreamWorker(PojoSessionContext sessionContext) {
-		return new StreamPojoWorkerImpl( indexedTypeManagers, sessionContext );
 	}
 
 	@Override
