+++ util/internal/integrationtest/common/src/main/java/org/hibernate/search/v6poc/util/impl/integrationtest/common/stub/backend/index/impl/StubChangesetIndexWorker.java
@@ -11,26 +11,55 @@ import java.util.List;
 import java.util.concurrent.CompletableFuture;
 
 import org.hibernate.search.v6poc.backend.index.spi.ChangesetIndexWorker;
+import org.hibernate.search.v6poc.backend.index.spi.DocumentContributor;
+import org.hibernate.search.v6poc.backend.index.spi.DocumentReferenceProvider;
 import org.hibernate.search.v6poc.engine.spi.SessionContext;
+import org.hibernate.search.v6poc.util.impl.integrationtest.common.stub.backend.document.StubDocumentNode;
 import org.hibernate.search.v6poc.util.impl.integrationtest.common.stub.backend.index.StubIndexWork;
 import org.hibernate.search.v6poc.util.impl.integrationtest.common.stub.backend.document.impl.StubDocumentElement;
 
-class StubChangesetIndexWorker extends StubIndexWorker
-		implements ChangesetIndexWorker<StubDocumentElement> {
+class StubChangesetIndexWorker implements ChangesetIndexWorker<StubDocumentElement> {
 	private final StubIndexManager indexManager;
+	private final SessionContext sessionContext;
 
 	private final List<StubIndexWork> works = new ArrayList<>();
 
 	private int preparedIndex = 0;
 
-	StubChangesetIndexWorker(StubIndexManager indexManager, SessionContext context) {
-		super( context );
+	StubChangesetIndexWorker(StubIndexManager indexManager, SessionContext sessionContext) {
+		this.sessionContext = sessionContext;
 		this.indexManager = indexManager;
 	}
 
 	@Override
-	protected void addWork(StubIndexWork work) {
-		works.add( work );
+	public void add(DocumentReferenceProvider documentReferenceProvider,
+			DocumentContributor<StubDocumentElement> documentContributor) {
+		StubIndexWork.Builder builder = StubIndexWork.builder( StubIndexWork.Type.ADD );
+		populate( builder, documentReferenceProvider );
+		StubDocumentNode.Builder documentBuilder = StubDocumentNode.document();
+		StubDocumentElement documentElement = new StubDocumentElement( documentBuilder );
+		documentContributor.contribute( documentElement );
+		builder.document( documentBuilder.build() );
+		addWork( builder.build() );
+	}
+
+	@Override
+	public void update(DocumentReferenceProvider documentReferenceProvider,
+			DocumentContributor<StubDocumentElement> documentContributor) {
+		StubIndexWork.Builder builder = StubIndexWork.builder( StubIndexWork.Type.UPDATE );
+		populate( builder, documentReferenceProvider );
+		StubDocumentNode.Builder documentBuilder = StubDocumentNode.document();
+		StubDocumentElement documentElement = new StubDocumentElement( documentBuilder );
+		documentContributor.contribute( documentElement );
+		builder.document( documentBuilder.build() );
+		addWork( builder.build() );
+	}
+
+	@Override
+	public void delete(DocumentReferenceProvider documentReferenceProvider) {
+		StubIndexWork.Builder builder = StubIndexWork.builder( StubIndexWork.Type.DELETE );
+		populate( builder, documentReferenceProvider );
+		addWork( builder.build() );
 	}
 
 	@Override
@@ -46,5 +75,15 @@ class StubChangesetIndexWorker extends StubIndexWorker
 		works.clear();
 		preparedIndex = 0;
 		return future;
+	}
+
+	private void populate(StubIndexWork.Builder builder, DocumentReferenceProvider documentReferenceProvider) {
+		builder.tenantIdentifier( sessionContext.getTenantIdentifier() );
+		builder.identifier( documentReferenceProvider.getIdentifier() );
+		builder.routingKey( documentReferenceProvider.getRoutingKey() );
+	}
+
+	private void addWork(StubIndexWork work) {
+		works.add( work );
 	}
 }
